refactor refactor event handling and repository methods for improved clarity and functionality
