+++ app/events/usecase/create_registration_event.go
@@ -11,7 +11,7 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-func (uc usecase) CreateRegisterEvent(ctx context.Context, payload domain.RegisterEventPayload) (domain.RegisterEventResponse, error) {
+func (uc usecase) CreateRegistrationEvent(ctx context.Context, payload domain.RegisterEventPayload) (domain.RegisterEventResponse, error) {
 	event, err := uc.repository.GetEvent(ctx, payload.EventID)
 	if err != nil {
 		logrus.Error("failed to get event")
@@ -47,7 +47,7 @@ func (uc usecase) CreateRegisterEvent(ctx context.Context, payload domain.Regist
 	}
 
 	err = uc.dbTX.StartTransaction(ctx, func(txCtx context.Context) error {
-		rId, err := uc.repository.CreateRegisterEvent(txCtx, domain.RegistrationEvent{
+		rId, err := uc.repository.CreateRegistrationEvent(txCtx, domain.RegistrationEvent{
 			OrderNo:     orderNo,
 			EventID:     event.ID,
 			Name:        payload.Name,
@@ -74,7 +74,7 @@ func (uc usecase) CreateRegisterEvent(ctx context.Context, payload domain.Regist
 				return err
 			}
 
-			_, err = uc.repository.CreatePayEvent(txCtx, domain.EventPay{
+			_, err = uc.repository.CreateEventPay(txCtx, domain.EventPay{
 				RegistrationEventID: rId,
 				EventID:             event.ID,
 				ImageProofPayment:   payload.ImageProofPayment,
